+++ teleop/htm/static/CSS/theme_mode.css
@@ -1,22 +1,285 @@
-body.dark-mode {
+/* Advanced mode */
+body.advanced-theme .advanced_view {
+  display: none !important;
+}
+
+
+/* #region Dark mode */
+body.dark-theme {
   background-color: var(--background_color_dark_mode);
   color: var(--text_color_dark_mode);
 }
 
-body.dark-mode .caret {
+body.dark-theme .caret {
   color: var(--text_color_dark_mode);
 }
 
-body.dark-mode #hamburger_menu_toggle .bar {
+body.dark-theme #hamburger_menu_toggle .bar {
   background-color: var(--background_color_light_mode);
 }
 
-body.dark-mode #video_stream_type {
+body.dark-theme #video_stream_type {
   background-color: var(--background_color_light_mode);
   color: var(--text_color_light_mode);
 }
 
+/* #endregion */
 
-body.advanced-mode .advanced_view {
-  display: none !important ;
+/* #region AI training feature */
+body.maneuver-training-feature .steeringWheel {
+  display: none;
 }
+
+body.maneuver-training-feature #horizontal_bar .current_mode_text:before {
+  content: "ai training"
+}
+
+body.maneuver-training-feature #mobile_controller_container #backward_square .trail_canvas,
+body.maneuver-training-feature #mobile_controller_container #backward_square .square_text {
+  display: none;
+}
+
+body.maneuver-training-feature #mobile_controller_container #forward_square .square_text:before {
+  content: 'forward';
+}
+
+body.maneuver-training-feature #mobile_controller_container #backward_square {
+  display: block;
+  width: 100%;
+  background: repeating-linear-gradient(-45deg, var(--background_color_light_mode), var(--background_color_light_mode) 16px, #ccc 18px, #ccc 11px);
+  border: 0;
+}
+
+body.maneuver-training-feature #mobile_controller_container .control_symbol {
+  display: none;
+}
+
+body.maneuver-training-feature .current_mode_button {
+  display: block;
+  background-color: #451c58;
+  color: white;
+  box-shadow: none;
+}
+
+body.maneuver-training-feature .current_mode_button:before {
+  content: "start"
+}
+
+/*  Training started state */
+body.maneuver-training-feature.training-started .current_mode_button {
+  background-color: #f41e52;
+  border: none;
+  content: 'stop';
+}
+
+body.maneuver-training-feature.training-started .current_mode_button:before {
+  content: "stop";
+}
+
+/* #endregion */
+
+/* #region Auto navigation feature */
+body.auto-navigation-feature .steeringWheel {
+  display: none;
+}
+
+body.auto-navigation-feature #horizontal_bar .current_mode_text:before {
+  content: "auto navigation"
+}
+
+body.auto-navigation-feature .current_mode_button:before {
+  content: 'start';
+}
+
+body.auto-navigation-feature #forward_square .square_text:before {
+  content: "Increase max speed"
+}
+
+body.auto-navigation-feature #backward_square .square_text:before {
+  content: "decrease max speed"
+}
+
+
+
+body.auto-navigation-feature .current_mode_button {
+  display: block;
+  background-color: #451c58;
+  color: white;
+  box-shadow: none;
+}
+
+body.auto-navigation-feature .current_mode_text,
+body.auto-navigation-feature #mobile_controller_container .trail_canvas {
+  display: none;
+}
+
+body.auto-navigation-feature .rover_speed,
+body.auto-navigation-feature .control_symbol {
+  display: flex;
+}
+
+
+body.auto-navigation-feature.navigation-started .current_mode_button {
+  background-color: #f41e52;
+  border: none;
+  content: 'stop';
+}
+
+body.auto-navigation-feature.navigation-started .current_mode_button:before {
+  content: 'stop';
+}
+
+/* #endregion */
+
+/* #region Confidence feature  */
+body.confidence-feature .current_mode_text {
+  display: block;
+}
+
+body.confidence-feature #horizontal_bar .current_mode_text:before {
+  content: "map recognize"
+}
+
+body.confidence-feature .steeringWheel {
+  display: none;
+}
+
+body.confidence-feature .current_mode_button {
+  display: block;
+  background-color: #451c58;
+  color: white;
+  box-shadow: none;
+}
+
+/* Recognize mode */
+body.confidence-feature.recognize-mode .current_mode_button:before {
+  content: 'start';
+}
+
+body.confidence-feature.recognize-mode .control_symbol,
+body.confidence-feature.recognize-mode .stop_text {
+  display: none;
+}
+
+/* Stop mode */
+body.confidence-feature.stop-mode .current_mode_button {
+  background-color: #f41e52;
+  border: none;
+}
+
+body.confidence-feature.stop-mode .current_mode_button:before {
+  content: 'stop';
+}
+
+body.confidence-feature.stop-mode #mobile_controller_container .square {
+  display: block;
+}
+
+body.confidence-feature.stop-mode .stop_text,
+body.confidence-feature.stop-mode .control_symbol {
+  display: none;
+}
+
+/* Return mode */
+body.confidence-feature.return-mode .current_mode_button {
+  background-color: #ffffff;
+  color: #451c58;
+}
+
+body.confidence-feature.return-mode .current_mode_button:before {
+  content: 'return';
+}
+
+/* Show result mode */
+body.confidence-feature.show-result-mode .current_mode_button {
+  display: block;
+  border: none;
+  content: 'Show result';
+}
+
+body.confidence-feature.show-result-mode .current_mode_button:before {
+  content: 'Show result';
+}
+
+/* Loading mode */
+body.confidence-feature.loading-mode .current_mode_state {
+  display: block;
+  color: #ff8a00;
+}
+
+body.confidence-feature.loading-mode .current_mode_state:before {
+  content: 'Loading...';
+}
+
+body.confidence-feature.loading-mode .current_mode_button {
+  display: none;
+}
+
+/* #endregion */
+
+/* #region Following feature */
+body.following-feature .current_mode_button {
+  display: block;
+}
+
+body.following-feature#horizontal_bar .current_mode_text:before {
+  content: "follow"
+}
+
+body.following-feature #mobile_controller_container .middle_section,
+body.following-feature #mobile_controller_container .square {
+  display: none;
+}
+
+/* Image mode */
+body.following-feature.image-mode .current_mode_button {
+  background-color: #ffffff;
+  border: none;
+}
+
+body.following-feature.image-mode .current_mode_button::before {
+  content: 'start ';
+}
+
+body.following-feature.image-mode #mobile_controller_container .current_mode_state,
+body.following-feature.image-mode #mobile_controller_container .square {
+  display: none;
+}
+
+/* Active mode */
+body.following-feature.active-mode .current_mode_button {
+  background-color: #f41e52;
+  border: none;
+}
+
+body.following-feature.active-mode .current_mode_button::before {
+  content: 'stop ';
+}
+
+body.following-feature.active-mode #mobile_controller_container .current_mode_state,
+body.following-feature.active-mode #mobile_controller_container .square {
+  display: none;
+}
+
+/* Inactive mode */
+body.following-feature.inactive-mode .current_mode_button {
+  background-color: #ffffff;
+  border: none;
+  content: 'start';
+}
+
+body.following-feature.inactive-mode #mobile_controller_container .square {
+  display: block;
+}
+
+/* Loading mode */
+body.following-feature.loading-mode #mobile_controller_container .current_mode_state:before {
+  content: 'Loading...';
+}
+
+body.following-feature.loading-mode .current_mode_button,
+body.following-feature.loading-mode #mobile_controller_container .square {
+  display: none;
+}
+
+/* #endregion */
