+++ teleop/htm/static/CSS/user_menu.css
@@ -132,7 +132,7 @@
   display: flex;
   align-items: center;
   text-decoration: none;
-  padding: 1rem 0;
+  margin: 1rem 0;
   transition: font-size 0.3s ease, color 0.3s ease;
 }
 
@@ -168,6 +168,12 @@
 
 .hamburger_menu_nav.active {
   left: 0;
+}
+#hamburger_menu_top_nav{
+  display: flex;
+  justify-content: flex-end;
+  align-items: flex-start;
+  flex-direction: column;
 }
 
 #hamburger_menu_bottom_nav {
