+++ teleop/htm/static/JS/Index/index_c_screen.js
@@ -1,6 +1,6 @@
 import CTRL_STAT from '../mobileController/mobileController_z_state.js'; // Stands for control state
 
-import { dev_tools, isMobileDevice, page_utils } from './index_a_utils.js';
+import { isMobileDevice, page_utils } from './index_a_utils.js';
 import { gamepad_controller } from './index_b_gamepad.js';
 import { gamepad_socket } from './index_e_teleop.js';
 
@@ -33,7 +33,7 @@ class DarkThemeManager {
 	}
 
 	setTheme(isDarkMode) {
-		this.body.classList.toggle('dark-mode', isDarkMode); //Add dark mode to body only
+		this.body.classList.toggle('dark-theme', isDarkMode); //Add dark mode to body only
 		this.updateLogo(isDarkMode);
 	}
 
@@ -90,9 +90,9 @@ class AdvancedThemeManager {
 
 	setAdvancedTheme() {
 		if (this.isAdvancedMode) {
-			$('body').removeClass('advanced-mode');
+			$('body').removeClass('advanced-theme');
 		} else {
-			$('body').addClass('advanced-mode');
+			$('body').addClass('advanced-theme');
 		}
 	}
 }
@@ -279,7 +279,6 @@ class MessageContainerManager {
 	}
 }
 
-
 class PathRenderer {
 	// Method to render a trapezoid shape
 	_renderTrapezoid(ctx, positions, fill = 'rgba(100, 217, 255, 0.3)') {
@@ -428,7 +427,6 @@ class RoverUI {
 				overlayLeftMarker1: this.getElement('div#overlay_left_marker1'),
 				overlayRightMarker0: this.getElement('div#overlay_right_marker0'),
 				overlayRightMarker1: this.getElement('div#overlay_right_marker1'),
-				infSpeed: this.getElement('div.inf_speed'),
 				autopilotOperatingTime: this.getElement('.inf_operating_time'),
 				messageBoxMessage: this.getElement('div#message_box_message'),
 				buttonTakeControl: this.getElement('input#message_box_button_take_control'),
@@ -445,7 +443,6 @@ class RoverUI {
 			this.overlayCenterMarkers = [elements.overlayCenterDistance0, elements.overlayCenterDistance1];
 			this.overlayLeftMarkers = [elements.overlayLeftMarker0, elements.overlayLeftMarker1];
 			this.overlayRightMarkers = [elements.overlayRightMarker0, elements.overlayRightMarker1];
-			this.elInfSpeed = elements.infSpeed;
 			this.elAutopilotOperatingTime = elements.autopilotOperatingTime;
 		} catch (error) {
 			console.error('Error in RoverUI.setNormalUIElements():', error);
@@ -579,13 +576,13 @@ class InferenceHandling {
 	}
 
 	updateAutopilotUI(infMessage) {
-		this.roverUI.elInfSpeed.show();
-		this.roverUI.elAutopilotOperatingTime.show();
+		$('div.inf_speed').show();
+		$('.inf_operating_time').show();
 		$('p.inf_speed_value').text(`${infMessage.max_speed.toFixed(1)} KM`);
 		$('div.inf_speed_label').text('Max Speed');
 		this.roverUI.renderDistanceIndicators('front');
 
-		this.controlCurrentModeButton(true);
+		// this.controlCurrentModeButton(true);
 
 		if (infMessage.ctl_activation > 0) {
 			this.updateAutopilotTimeDisplay(infMessage.ctl_activation);
@@ -593,11 +590,9 @@ class InferenceHandling {
 	}
 
 	updateAutopilotTimeDisplay(ctlActivation) {
-		const elAutopilotOperatingTime = $('.inf_operating_time');
 		const time = this.formatTime(ctlActivation * 1e-3); // Convert ms to seconds
-
-		elAutopilotOperatingTime.text(time);
-		elAutopilotOperatingTime.css('color', 'rgb(100, 217, 255)');
+		$('.inf_operating_time').text(time);
+		$('.inf_operating_time').css('color', 'rgb(100, 217, 255)');
 	}
 
 	formatTime(totalSeconds) {
@@ -613,13 +608,11 @@ class InferenceHandling {
 	}
 
 	hideAutopilotUI() {
-		const elAutopilotOperatingTime = this.roverUI.elAutopilotOperatingTime;
+		$('div.inf_speed').hide();
+		$('.inf_operating_time').text('00:00:00');
+		$('.inf_operating_time').hide();
 
-		this.roverUI.elInfSpeed.hide();
-		elAutopilotOperatingTime.text('00:00:00');
-		elAutopilotOperatingTime.hide();
-
-		this.controlCurrentModeButton(false);
+		// this.controlCurrentModeButton(false);
 	}
 
 	updateSteeringWheelRotation(speed, steeringAngle) {
@@ -670,8 +663,8 @@ class CameraControls {
 
 		// Add listeners to update RoverUI whenever the camera changes
 		this.addCameraActivationListener((name) => {
-			this.roverUI.activeCamera = name; 
-			this.roverUI.renderDistanceIndicators(name); 
+			this.roverUI.activeCamera = name;
+			this.roverUI.renderDistanceIndicators(name);
 		});
 	}
 	selectNextCamera() {
@@ -758,4 +751,4 @@ const roverUI = new RoverUI();
 const inferenceHandling = new InferenceHandling(roverUI);
 const cameraControls = new CameraControls(roverUI);
 
-export {  roverUI, inferenceHandling, cameraControls, helpMessageManager, messageContainerManager, advancedThemeManager };
+export { advancedThemeManager, cameraControls, helpMessageManager, inferenceHandling, messageContainerManager, roverUI };
