+++ teleop/htm/templates/index.html
@@ -5,14 +5,15 @@
 		<meta http-equiv="X-UA-Compatible" content="IE=edge" />
 		<meta name="viewport" content="width=device-width, initial-scale=1.0" />
 		<title>teleop</title>
-		<!-- dark mode should be the highest to overcome all of the other styles -->
-		<link rel="stylesheet" type="text/css" href="../static/CSS/dark_mode.css" />
+
 		<link rel="stylesheet" type="text/css" href="../static/CSS/style.css" />
 		<link rel="stylesheet" type="text/css" href="../static/CSS/user_menu.css" />
 		<link rel="stylesheet" type="text/css" href="../static/CSS/menu_controls.css" />
 		<link rel="stylesheet" type="text/css" href="../static/CSS/menu_logbox.css" />
 		<link rel="stylesheet" type="text/css" href="../static/CSS/menu_settings.css" />
 		<link rel="stylesheet" type="text/css" href="../static/CSS/mobileController.css" />
+		<!-- dark mode should be the highest to overcome all of the other styles -->
+		<link rel="stylesheet" type="text/css" href="../static/CSS/theme_mode.css" />
 		<link rel="stylesheet" href="../static/external/jquery-ui-1.12.1.min.css" />
 		<link rel="shortcut icon" href="#" />
 	</head>
