+++ teleop/htm/templates/mobile_controller_ui.html
@@ -1,7 +1,7 @@
 <div id="mobile_controller_container">
 	<div id="horizontal_bar">
 		<div class="left_section">
-			<p class="current_mode_text">manual drive</p>
+			<p class="current_mode_text"></p>
 			<div class="inf_speed speed_value">
 				<p class="inf_speed_value">0</p>
 				<div class="inf_speed_label"></div>
@@ -10,10 +10,10 @@
 		</div>
 		<div class="middle_section">
 			<img src="../static/assets/manual_mode.png" class="steeringWheel deep_purple" />
-			<span class="autopilot_status"></span>
+
 		</div>
 		<div class="right_section">
-			<button class="current_mode_button">START FOLLOWING</button>
+			<button class="current_mode_button"></button>
 			<!-- The text inside `current mode button` is just a place holder. It will change with each mode -->
 			<p class="current_mode_state"></p>
 		</div>
@@ -24,14 +24,14 @@
 
 		<div id="forward_square" class="square">
 			<canvas class="trail_canvas"></canvas>
-			<div class="square_text">forward</div>
+			<p class="square_text"></p>
 			<span class="stop_text">drag your finger to this area to force the robot to stop</span>
 			<div class="control_symbol plus"></div>
 		</div>
 
 		<div id="backward_square" class="square">
 			<canvas class="trail_canvas"></canvas>
-			<div class="square_text">backward</div>
+			<p class="square_text"></p>
 			<span class="stop_text">drag your finger to this area to force the robot to stop</span>
 			<div class="control_symbol minus"></div>
 		</div>
