+++ teleop/htm/templates/normal_ui.html
@@ -48,9 +48,9 @@
 
 	<!-- Middle Section -->
 	<div class="middle_section">
-		<p class="current_mode_text">manual drive</p>
+		<p class="current_mode_text"></p>
 		<img src="../static/assets/manual_mode.png" class="steeringWheel deep_purple current_mode_img" />
-		<button class="current_mode_button">START FOLLOWING</button>
+		<button class="current_mode_button"></button>
 		<p class="current_mode_state"></p>
 		<div class="inf_speed speed_value">
 			<p class="inf_speed_value">0</p>
